requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2
python-multipart==0.0.9
email-validator==2.2.0
pytest==8.2.0
pytest-asyncio==0.24.0
pytest-cov==4.1.0
Pillow==10.4.0
